requirements.txt
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
pymongo==3.10.1
python-dotenv==0.13.0